backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==47.0.0
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.29.0
fsspec==2026.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.13.0
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
orjson==3.11.8
packaging==26.2
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pycparser==3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.20.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.27
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
rich==15.0.0
rich-toolkit==0.19.7
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.37.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
tqdm==4.67.3
transformers==5.7.0
typer==0.25.1
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
ujson==5.12.0
uvicorn==0.46.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.7.0
websockets==16.0
wrapt==2.1.2
